feat: add subscription permissions

+++ src/GR.Extensions/GR.Ecommerce.Extension/GR.ECommerce.Abstractions/Models/ProductVariation.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 using GR.Core;
 
 namespace GR.ECommerce.Abstractions.Models
@@ -10,7 +9,7 @@ namespace GR.ECommerce.Abstractions.Models
         public virtual Product Product { get; set; }
         public virtual Guid ProductId { get; set; }
 
-        public virtual double Price { get; set; } = 0;
+        public virtual decimal Price { get; set; } = 0;
 
         public virtual List<ProductVariationDetail> ProductVariationDetails { get; set; }
